HBSD: Add MK_UNINIT_AUTOINIT feature
In an effort to auto-initialize stack-based variables, add a
MK_UNINIT_AUTOINIT feature, default to "no". Only the userland portion
is included in this commit. Further work research and development on the
kernel side is being performed.

Note that inclusion of this feature does not endorse it. Inclusion is
meant to provide another avenue for research on efficacy and efficiency.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>