HBSD: Enable PaX SEGVGUARD by default
When PAX_HARDENING is set, enable PaX SEGVGUARD by default. Though the
implementation is not as performant as desired, I believe it to be
sufficient for average workloads.

Users wanting to squeeze every last bit of juice out of their system
would want to switch PaX SEGVGUARD back to opt-in mode
(`sysctl hardening.pax.segvguard.status=1`).

Developers who would like to help improve the performance and accuracy
of this SEGVGUARD implementation are encouraged to submit a Pull
Request.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
X-MFC-able:	no