HBSD: Fix ASR logic
The do_asr and et_dyn_addr variables weren't being properly set after
the merge of FreeBSD ASR into HardenedBSD. Under some conditions, do_asr
could be set to 1. Part of the problem stems from kib@freebsd's abuse of
the et_dyn_addr variable, using it as a bool in some cases and an
address in others. We introduced the do_asr variable to undo that
insanity, but the logic still wasn't quite right.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>