HBSD: Fix libarchive's handling of extattr_set_fd
On FreeBSD, extattr_set_fd does not return the same as extattr_set_file.
It returns zero on success, non-zero on failure.

We can detect the failure by manually setting errno prior to the call
and checking after.

If errno remains zero, fake the return value by setting e to size.

This is a hack for now until I (Shawn Webb) get FreeBSD to fix the
issue, if that's even possible.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
issue:		#1
MFC-to:		12-STABLE