HBSD: When jailed, do not attempt to load zfs kernel module
libzfs tries to be sneaky and auto-load the zfs kernel module if it's
not already loaded. This isn't possible to do in a jailed context. In
fact, jailed processes cannot even use the kernel module interface, not
even to list kernel modules.

This commit teaches libzfs to not look for and auto-load the zfs kernel
module if the process is running under a jailed context.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
Reported-by:	Hinnerk van Bruinehsen
issue:		#31
MFC-to:		13-STABLE
MFC-to:		12-STABLE