HBSD: Hack llvm-ar to exibit the same behavior on error as elftc-ar
elftc-ar doesn't exit with a non-zero exitcode on an error condition.
The ports build framework gets tricked by elftc-ar in these error
conditions. llvm-ar does the right thing, but that causes some ports[0]
to fail to build. For now, introduce a hack in llvm-ar to match the
(what I think is incorrect) behavior in elftc-ar.

From here, I'll report on the freebsd-toolchain mailing list, where
there's already an experimental patch for using llvm-ar, llvm-nm, and
llvm-objdump for building base. I'll also report to elftc-ar about the
errant error handling. We'll keep the patch until either elftc-ar or
the majority of ports entries are fixed.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>