HBSD: Disable sanity check in llvm-objcopy
The sanity check for ensuring which symbols to remove (when llvm-objcopy
is running as the strip command) is a bit too aggressive. Relax the
rules, at least temporarily, to address at least some of the ports build
failures[0].

There's more work to be done, and ideally, we'd re-enable the sanity
check.

[0]: http://ci-08.md.hardenedbsd.org/data/hardenedbsd-current_amd64-local/2021-10-08_15h25m44s/logs/errors/postgresql12-client-12.8.log

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
issue:		#66