HBSD: Fix copy error for the bootstrap
Following the commit "HBSD: Apply the kernel update first", the "-r" bootstrap option was broken.
Indeed the "apply_mtree" function is not executed before the "apply_kernel"
which has the effect of breaking the kernel copy in the bootstrap.

Although several solutions are possible,
I prefer to force the creation of the kernel directory so that the "cp" command does not fail.

Signed-off-by: Loic <loic.f@hardenedbsd.org>
Reported-by: Jonas Rinner @Follpvosten
Issue:		#55
MFC-to:		13-STABLE
MFC-to:		12-STABLE