HBSD: Resolve merge conflicts
Usually, the merge conflict resolution commit is self-explanitory. This
commit, however, needs extra comments.

In FreeBSD commit 706f4a81a81250a326ea25914e7effe1768f1a37, FreeBSD
introduced a helper macro for getting the (new to that commit)
per-process psstrings. FreeBSD did this to support randomizing the stack
top address.

The macro itself is useful, though the FreeBSD version of it points to
their version of per-process psstrings. I've changed the macro to use
HardenedBSD's version (proc->p_psstrings). Those places in the kernel
that conflicted with our per-process psstrings and theirs, I switched to
using the new macro. There's a few places where merge conflicts did not
happen and that still should be changed (though leaving them as-is is
harmless).

To simplify life for HardenedBSD, I've removed a good chunk of the
FreeBSD AS{L}R support in the ELF image activator. This drastically
simplifies the code, and helps prevent potential bugs. There's a lot of
other places which need the same type of removal/simplification.

FreeBSD plans to merge into 13-STABLE their new stack randomization
work in two weeks. Prior to that, I plan to finish the work to migrate
us to the new per-process psstrings macro and to remove all traces of
FreeBSD's AS{L}R implementation. I'll merge that work into 13-STABLE
prior to resolving the inevitable merge conflict from FreeBSD's MFC.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>