HBSD: Properly prevent process tracing in OpenSSH
When procctl is passed getpid(), it performs an additional check to see
whether the process can be seen or debugged. In the case of the sftpd
server, the sftpd server switches UID/GID to that of the to-be-logged-in
user account.

Since HardenedBSD sets the security.bsd.unprivileged_proc_debug sysctl
node to 0 by default (FreeBSD leaves it at 1), procctl's cansee/candebug
check fails. This cascades down to the sftpd process, causing sftpd to
deny the connection.

To circumvent the cansee/candebug check, procctl callers can pass in 0
as the target PID, which tells procctl to act on the caller itself. As
such, the cansee/candebug check is bypassed, since it's okay to use
procctl on oneself.

To fix this problem, have OpenSSH pass in 0 as the target PID.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>