HBSD: Fix kernel-side of procctl(2) cansee/candebug generally
When a non-zero PID value is passed to procctl(2), procctl verifies that
the caller has the right to observe the target. We should treat passing
in the caller's own PID as the same as passing in 0, in which the
cansee/candebug check is bypassed.

This fixes in a more general fashion the bug that OpenSSH manifest,
which I had fixed in commit cc791ab8ae6b6e29fe26ac6cf247d0a478eef4cb.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>