HBSD: Partially revert FreeBSD commit argc == 0 commit
FreeBSD implements the argc check too early in the execve stack. A
system entry vector's sv_copyout_strings may, in theory, modify the
to-be-executed executable image's argument list, and cause argc to be 0.
FreeBSD's check is insufficient in that regard.

As such, favor HardenedBSD's implementation, which does take into
account late-stage modification of argc.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>