HBSD: Expose CFI-safe wrappers for dbopen(3)-related function pointers
The dbopen(3) API requires downstream consumers to directly reference
and call function pointers in a contextual structure. This causes issues
with the cfi-icall scheme specifically since the function pointers in
question point to uninstrumented functions in libc.

So that applications can still apply the cfi-icall scheme, we need to
provide simple wrappers for the in-ABI function pointers. The wrappers
are added to the libc symbol map such that applications, like sa(8), can
make use of them.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
issue:		#75
MFC-to:		13-STABLE