HBSD: Disallow chroot(2) when a directory is opened
When PAX_HARDENING is enabled, set the sysctl knob that prevents
successful chroot(2) when one or more directories are opened by the
calling process.

Determining which applications in ports that violate this policy would
be a useful academic adventure. Even more those applications which do
not check the return value of chroot(2).

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
MFC-to:		13-STABLE