HBSD: Harden object tracing in the RTLD
The RTLD provides a capability to trace which shared objects an object
file depends on, which is what ldd(1) uses behind-the-scenes.

When we're in hardened RTLD mode, disable the shared object tracing
functionality. This can help guard against information leak attacks.
This comes at a cost of an ldd(1) that displays no meaningful output.

In order to make ldd(1) meaningful again, the `hardening.harden_rtld`
sysctl node would need to be set to 0 first.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
MFC-to:		13-STABLE