HBSD: Randomize default packet TTL value at boot
When the new HBSD_RESIST_FINGERPRINTING kernel option is enabled (as it
is on amd64 with this commit), set the default packet TTL to a random
value between 32 and 255.

Rondomizing the default TTL value at boot time can help with regards to
resisting fingerprinting attacks and preventing information leaks. An

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
MFC-to:		13-STABLE