HBSD: Fully prohibit the RTLD from using a memory-backed file descriptor
This fully mitigates abuse of the RTLD to load a memory-backed shared
object. As it stands today, the shared memory subsystem does not support
fstatfs(2). When fstatfs(2) is called on a file descriptor created by
shm_open(2)/__sys_shm_open2/memfd_create(2), fstatfs(2) will fail and
set errno to an undocumented EINVAL.

We can be "clever" by refusing to load any file descriptor that fails
the fstatfs(2) check.

Signed-off-by:	Shawn Webb <shawn.webb@hardenedbsd.org>
MFC-to:		13-STABLE